Fold Gemma4 inference and pretrained modules into model.py
Match the Whisper template, where the whole model (bundle factory, step
models, and weight loading) lives in one model.py. Gemma4 previously split
this across model.py (backbone), inference.py (cached step models), and
pretrained.py (Gemma4 bundle + weight download/loading), which forced the odd
`pretrained` module and a model <- inference <- pretrained import layering.

Merging all three into model.py removes the cross-module imports entirely, so
the layering (and its latent circular-import risk) disappears, and Gemma4 is
now imported the same way as Whisper:
`from paz.models.foundation.gemma4.model import Gemma4`.

- Move Gemma4DecoderStep / Gemma4MultimodalDecoderStep / PerLayerEmbeddingStep,
  build_empty_cache, and the Gemma4 bundle factory + weight download/sharding
  helpers into model.py; delete inference.py and pretrained.py.
- Repoint every importer (applications, decoding, multimodal_decoding,
  conversion, models/__init__, the demo, and tests) to gemma4.model.
- Fold inference_test.py into model_test.py.
