+++ paz/models/__init__.py
@@ -34,4 +34,4 @@ from .foundation.dinov3.models.vision_transformer import DINOV3VITS
 from .foundation.dinov3.models.vision_transformer import DINOV3VITB
 from .foundation.dinov3.models.vision_transformer import DINOV3VITL
 from .foundation.whisper.model import Whisper
-from .foundation.gemma4.pretrained import Gemma4
+from .foundation.gemma4.model import Gemma4
