+++ paz/models/foundation/gemma4/gemma4_test.py
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 
 from paz.models import Gemma4
-from paz.models.foundation.gemma4 import pretrained
+from paz.models.foundation.gemma4 import model
 from paz.models.foundation.gemma4.tokenizer import Gemma4Tokenizer
 from paz.models.foundation.gemma4.multimodal_decoding import generate_eager
 
@@ -26,15 +26,15 @@ def test_shard_and_reassemble_round_trips(tmp_path):
     blob = os.urandom(5_000_001)
     source = tmp_path / "decoder_step.weights.h5"
     source.write_bytes(blob)
-    parts = pretrained.split_file(source, tmp_path, "gemma4_2b_decoder",
+    parts = model.split_file(source, tmp_path, "gemma4_2b_decoder",
                                   part_bytes=2_000_000)
     assert parts == ["gemma4_2b_decoder.part0", "gemma4_2b_decoder.part1",
                      "gemma4_2b_decoder.part2"]
     part_paths = [tmp_path / name for name in parts]
     output = tmp_path / "reassembled.h5"
-    pretrained.concatenate_parts(part_paths, output)
+    model.concatenate_parts(part_paths, output)
     assert output.read_bytes() == blob
-    assert pretrained.compute_sha256(output) == pretrained.compute_sha256(source)
+    assert model.compute_sha256(output) == model.compute_sha256(source)
 
 
 def test_shard_weights_writes_manifest(tmp_path):
@@ -43,7 +43,7 @@ def test_shard_weights_writes_manifest(tmp_path):
     (source_dir / "config.json").write_text("{}")
     (source_dir / "decoder_step.weights.h5").write_bytes(os.urandom(3_000_000))
     output_dir = tmp_path / "release"
-    manifest_path = pretrained.shard_weights(
+    manifest_path = model.shard_weights(
         source_dir, "gemma4_2b", output_dir, part_bytes=2_000_000)
     manifest = json.loads(manifest_path.read_text())
     assert manifest["config.json"]["parts"] == ["gemma4_2b_config.json.part0"]
@@ -52,18 +52,18 @@ def test_shard_weights_writes_manifest(tmp_path):
     decoder = manifest["decoder_step.weights.h5"]
     paths = [output_dir / asset for asset in decoder["parts"]]
     merged = output_dir / "merged.h5"
-    pretrained.concatenate_parts(paths, merged)
-    assert pretrained.compute_sha256(merged) == decoder["sha256"]
+    model.concatenate_parts(paths, merged)
+    assert model.compute_sha256(merged) == decoder["sha256"]
 
 
 def test_assemble_detects_corruption(tmp_path, monkeypatch):
     (tmp_path / "part0").write_bytes(os.urandom(2048))
     monkeypatch.setattr(
-        pretrained, "get_file",
+        model, "get_file",
         lambda asset, url, cache_subdir: str(tmp_path / "part0"))
     entry = {"parts": ["part0"], "sha256": "0" * 64}
     with pytest.raises(ValueError):
-        pretrained.assemble_weights_file(tmp_path / "out.h5", entry, "sub")
+        model.assemble_weights_file(tmp_path / "out.h5", entry, "sub")
 
 
 @pytest.mark.skipif(not run_weights_test(), reason="set GEMMA4_WEIGHTS_TEST=1")
